Fix cp_reflink behaviour
Unfortunately, "sys.platform == 'linux2'" is a bad predictor of reflink
support. It depends on the availability of the GNU toolchain (not always the
case, e.g. on OS X), if a recent enough kernel is running and if the underlying
file system supports it. So we just try with reflink and fall back to a
traditional cp if the first attempt failed.

Re #13328