don't clone backend revisions for force-full backup
This is quick fix to avoid the full backup to read all existing data, while a mere checksum comparison would be sufficient. It currently makes only sense with chunked backend.

A more general approach would be to put more logic into the chunks, so they only read data when it's really necessary.