daemon: fix iteration race condition
We've encountered a `dictionary changed size during iteration`
RuntimeError in the wild for the jobs dictionary. This is likely the
result of an asynchronous race condition, where another coroutine had
modified the dictionary while the the iteration was async-interrupted.

We instead now follow the best practice of iterating on a copy of the
values. Having stale values around for status calculation is no issue,
as the changes will just be included in the next status update instead.

I've also checked all other iterations of `self.jobs` in the daemon for
similar issues, found none.