use the same parent revision in source and backend
The parent of a revision was previously selected independently of
additional source constraints. This could result in a revision C having
the parent B, but with a delta of A -> C applied by the source (e.g.
the snapshot for B is missing).
This did not cause any issues until now because ceph always includes
the chunks of the delta B->C in A->C (even when A->B is the inverse of
B->C), but we should not rely on this behaviour.

The file source also gained the ability to select a trusted parent
instead of falling back to a non-cow copy when a distrusted revision
exists.