Coordinate backups for the same job between servers
The server with the largest number of local revisions is the leader.
If the leader is offline another server will take over.

This assumes that all servers share the same rng and thus
schedule the backup at the same time. If this is not the case more
backups than necessary may be created (but not less).